+++ src/components/dialog/index.tsx
@@ -1,7 +1,14 @@
 import { View } from '@tarojs/components';
 import { ViewProps } from '@tarojs/components/types/View';
 import classNames from 'classnames';
-import { ReactNode, FC } from 'react';
+import {
+  forwardRef,
+  useImperativeHandle,
+  ReactNode,
+  FC,
+  useState,
+  cloneElement,
+} from 'react';
 import Typography from '../typography';
 import Button from '../button';
 import Flex from '../flex';
@@ -12,11 +19,23 @@ export type P = {
   visible?: boolean;
   title?: string;
   showCancelButton?: boolean;
-  onConfirm?: () => void;
+  theme?: 'radio';
+  closeOnMaskClick?: boolean;
+  content?: ReactNode;
+  message?: string;
+  footer?: ReactNode;
+  onConfirm?: () => void | Promise<any>;
   onCancel?: () => void;
 };
 
-const Dialog: FC<P & Omit<ViewProps, 'onClick'>> = ({
+export type DialogInstance = {
+  show: () => void;
+  hide: () => void;
+};
+
+const Dialog: FC<P & Omit<ViewProps, 'onClick'>> & {
+  Alert: any;
+} = ({
   children,
   visible,
   title,
@@ -24,8 +43,15 @@ const Dialog: FC<P & Omit<ViewProps, 'onClick'>> = ({
   onConfirm,
   onCancel,
   className,
+  message,
+  theme,
+  closeOnMaskClick,
+  content,
+  footer,
   ...props
 }) => {
+  const [confireLoading, setConfireLoading] = useState(false);
+  const [cancelLoading, setCancelLoading] = useState(false);
   const prefixCls = 'tv-dialog';
   const classes = classNames(
     prefixCls,
@@ -34,51 +60,180 @@ const Dialog: FC<P & Omit<ViewProps, 'onClick'>> = ({
     },
     className,
   );
+  const onConfirmBtn = () => {
+    if (onConfirm) {
+      const target: any = onConfirm();
+      if (target && typeof target.then === 'function') {
+        setConfireLoading(true);
+        try {
+          target.finally(() => setConfireLoading(false));
+        } catch (err) {
+          console.log(err);
+          setConfireLoading(false);
+        }
+      }
+    }
+  };
+  const onCancelBtn = () => {
+    if (onCancel) {
+      const target: any = onCancel();
+      if (target && typeof target.then === 'function') {
+        setCancelLoading(true);
+        try {
+          target.finally(() => setCancelLoading(false));
+        } catch (err) {
+          console.log(err);
+          setCancelLoading(false);
+        }
+      }
+    }
+  };
+  const confirmProps: any = {
+    onClick: onConfirmBtn,
+    loading: confireLoading,
+    type: 'primary',
+  };
+  const cancelProps: any = {
+    onClick: onCancelBtn,
+    loading: cancelLoading,
+    plain: true,
+  };
+  /** 按钮底部普通按钮 */
+  const FooterBaseItem = () => {
+    return (
+      <>
+        {showCancelButton ? (
+          <>
+            <Button
+              {...cancelProps}
+              className={`${prefixCls}-footer--btn ${prefixCls}-footer--btn--more`}
+            >
+              取消
+            </Button>
+            <Button
+              {...confirmProps}
+              plain
+              className={`${prefixCls}-footer--btn ${prefixCls}-footer--btn--more`}
+            >
+              确定
+            </Button>
+          </>
+        ) : (
+          <Button
+            {...confirmProps}
+            block
+            plain
+            className={`${prefixCls}-footer--btn`}
+          >
+            确定
+          </Button>
+        )}
+      </>
+    );
+  };
+
+  /** 按钮底部圆形按钮 */
+  const FooterRadioItem = () => {
+    return (
+      <>
+        {showCancelButton ? (
+          <Flex className={`${prefixCls}-footer--radio`} justify='center'>
+            <Button
+              {...cancelProps}
+              round
+              className={`${prefixCls}-footer--radio__btn`}
+            >
+              取消
+            </Button>
+            <Button
+              {...confirmProps}
+              round
+              className={`${prefixCls}-footer--radio__btn`}
+            >
+              确定
+            </Button>
+          </Flex>
+        ) : (
+          <Button
+            {...confirmProps}
+            block
+            plain
+            className={`${prefixCls}-footer--btn`}
+          >
+            确定
+          </Button>
+        )}
+      </>
+    );
+  };
   return (
-    <Overlay visible={visible} {...props}>
+    <Overlay
+      onClick={() => closeOnMaskClick && onCancel && onCancel()}
+      visible={visible}
+      {...props}
+    >
       <View className={`${prefixCls}-body`}>
         <View className={classes} onClick={(e) => e.stopPropagation()}>
-          <Typography.Title className={`${prefixCls}-head`} align='center'>
-            {title}
-          </Typography.Title>
-          <Flex className={`${prefixCls}-content`} justify='center'>
-            {children}
-          </Flex>
-          <Flex className={`${prefixCls}-footer`} wrap='nowrap'>
-            {showCancelButton ? (
-              <>
-                <Button
-                  onClick={onCancel}
-                  plain
-                  className={`${prefixCls}-footer--btn ${prefixCls}-footer--btn--more`}
-                >
-                  取消
-                </Button>
-                <Button
-                  onClick={onConfirm}
-                  plain
-                  type='primary'
-                  className={`${prefixCls}-footer--btn ${prefixCls}-footer--btn--more`}
-                >
-                  确定
-                </Button>
-              </>
-            ) : (
-              <Button
-                block
-                onClick={onConfirm}
-                plain
-                type='primary'
-                className={`${prefixCls}-footer--btn`}
-              >
-                确定
-              </Button>
-            )}
-          </Flex>
+          {title && (
+            <Typography.Title className={`${prefixCls}-head`} align='center'>
+              {title}
+            </Typography.Title>
+          )}
+          {content || (
+            <Flex className={`${prefixCls}-content`} justify='center'>
+              {message ? (
+                <Typography.Text>{message}</Typography.Text>
+              ) : (
+                children
+              )}
+            </Flex>
+          )}
+          {footer || (
+            <Flex className={`${prefixCls}-footer`} wrap='nowrap'>
+              {theme === 'radio' ? <FooterRadioItem /> : <FooterBaseItem />}
+            </Flex>
+          )}
         </View>
       </View>
     </Overlay>
   );
 };
+
+const Alert = forwardRef<
+  DialogInstance,
+  Omit<P, 'visible'> & Omit<ViewProps, 'onClick'>
+>(({ children, onCancel, onConfirm, ...props }, ref) => {
+  const [visible, setVisible] = useState(false);
+  useImperativeHandle(ref, () => ({
+    show: () => setVisible(true),
+    hide: () => setVisible(false),
+  }));
+  const onAlterCancel = () => {
+    setVisible(false);
+    onCancel && onCancel();
+  };
+  const onAlterConfirm = () => {
+    setVisible(false);
+    onConfirm && onConfirm();
+  };
+  return (
+    <>
+      {children &&
+        cloneElement(children as any, {
+          onClick: () => {
+            setVisible(true);
+          },
+        })}
+      <Dialog
+        onCancel={onAlterCancel}
+        onConfirm={onAlterConfirm}
+        visible={visible}
+        {...props}
+      />
+    </>
+  );
+});
+
+Dialog.Alert = Alert;
 
 export default Dialog;